+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 
 # SnakeCharm Plugin Changelog
 
+## [Unreleased]
+Released on ...
+                                          
+### Fixed 
+- "Snakemake support is disabled. Not all features are available" is show in non Snakemake context  (see [#491](https://github.com/JetBrains-Research/snakecharm/issues/491))
+- TODO  (see [#NNN](https://github.com/JetBrains-Research/snakecharm/issues/NNN)) 
+
 ## [2022.1.743]
 Released on May 4th 2022